# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Talent Management Skills API",
    "description": "API for querying employee skills, proficiency levels, evidence, skill coverage, talent search, and organizational skill analytics. Built on a PostgreSQL talent management database with hand-written SQL queries.",
    "version": "1.0.0",
    "contact": {
      "name": "TM Skills API"
    }
  },
  "servers": [
    {
      "url": "https://tm-skills-api.cfapps.ap10.hana.ondemand.com",
      "description": "Production (SAP BTP Cloud Foundry)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/tm/employees/{employee_id}/skills": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Get employee skill profile",
        "description": "What skills does this employee have, at what proficiency and confidence?",
        "operationId": "getEmployeeSkills",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "description": "Employee ID (e.g. EMP000001)",
            "schema": {
              "type": "string",
              "pattern": "^EMP\\d{6}$",
              "example": "EMP000001"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee skill profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeSkillProfile"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/tm/employees/{employee_id}/skills/{skill_id}/evidence": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Get skill evidence for employee",
        "description": "Why do we think this employee is proficient in this skill? Returns all evidence items.",
        "operationId": "getEmployeeSkillEvidence",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "description": "Employee ID (e.g. EMP000001)",
            "schema": {
              "type": "string",
              "pattern": "^EMP\\d{6}$",
              "example": "EMP000001"
            }
          },
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "description": "Skill ID",
            "schema": {
              "type": "number",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Skill evidence for the employee",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillEvidenceResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/tm/employees/{employee_id}/top-skills": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Get employee's strongest skills",
        "description": "Skill passport: what are this employee's strongest skills?",
        "operationId": "getEmployeeTopSkills",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "description": "Employee ID (e.g. EMP000001)",
            "schema": {
              "type": "string",
              "pattern": "^EMP\\d{6}$",
              "example": "EMP000001"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of top skills to return",
            "schema": {
              "type": "number",
              "default": 10,
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee's top skills",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeTopSkills"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/tm/employees/{employee_id}/evidence": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Get all evidence for employee",
        "description": "All evidence items across all skills for this employee.",
        "operationId": "getEmployeeEvidence",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "description": "Employee ID (e.g. EMP000001)",
            "schema": {
              "type": "string",
              "pattern": "^EMP\\d{6}$",
              "example": "EMP000001"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All evidence for the employee",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeEvidenceInventory"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/tm/orgs/{org_unit_id}/skills/summary": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get org unit skill summary",
        "description": "What are the top skills in this org unit (including child orgs)?",
        "operationId": "getOrgSkillSummary",
        "parameters": [
          {
            "name": "org_unit_id",
            "in": "path",
            "required": true,
            "description": "Org unit ID (e.g. ORG030 or ORG031B)",
            "schema": {
              "type": "string",
              "pattern": "^ORG\\d{1,4}[A-Z]?$",
              "example": "ORG030"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of top skills to return",
            "schema": {
              "type": "number",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Org skill summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgSkillSummaryResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/tm/orgs/{org_unit_id}/skills/{skill_id}/experts": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get skill experts within an org unit",
        "description": "Who in this org unit has this skill at minimum proficiency?",
        "operationId": "getOrgSkillExperts",
        "parameters": [
          {
            "name": "org_unit_id",
            "in": "path",
            "required": true,
            "description": "Org unit ID (e.g. ORG030 or ORG031B)",
            "schema": {
              "type": "string",
              "pattern": "^ORG\\d{1,4}[A-Z]?$",
              "example": "ORG030"
            }
          },
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "description": "Skill ID",
            "schema": {
              "type": "number",
              "example": 1
            }
          },
          {
            "name": "min_proficiency",
            "in": "query",
            "required": false,
            "description": "Minimum proficiency level",
            "schema": {
              "type": "number",
              "default": 3,
              "minimum": 0,
              "maximum": 5
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Max results",
            "schema": {
              "type": "number",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Skill experts in the org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgSkillExpertsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Pass in the X-API-Key header."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "detail": "Missing API key \u2014 include an X-API-Key header"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "detail": "Invalid API key"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "detail": "Employee not found"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "detail"
        ]
      },
      "EmployeeRef": {
        "type": "object",
        "description": "Basic employee reference information",
        "properties": {
          "employee_id": {
            "type": "string",
            "description": "Employee ID",
            "example": "EMP000001"
          },
          "display_name": {
            "type": "string",
            "nullable": true,
            "description": "Employee display name"
          },
          "job_title": {
            "type": "string",
            "nullable": true,
            "description": "Current job title"
          },
          "job_family": {
            "type": "string",
            "nullable": true,
            "description": "Job family"
          },
          "org_name": {
            "type": "string",
            "nullable": true,
            "description": "Organization name"
          },
          "seniority_level": {
            "type": "integer",
            "nullable": true,
            "description": "Seniority level"
          }
        },
        "required": [
          "employee_id"
        ]
      },
...
      "EmployeeSkillEntry": {
        "type": "object",
        "description": "A single skill entry for an employee",
        "properties": {
          "skill_id": {
            "type": "integer",
            "description": "Skill ID"
          },
          "skill_name": {
            "type": "string",
            "description": "Skill name"
          },
          "category": {
            "type": "string",
            "description": "Skill category"
          },
          "proficiency": {
            "type": "integer",
            "description": "Proficiency level (0-5)"
          },
          "confidence": {
            "type": "integer",
            "description": "Confidence score"
          },
          "source": {
            "type": "string",
            "description": "Source of the skill assessment"
          },
          "last_updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the skill was last updated"
          }
        },
        "required": [
          "skill_id",
          "skill_name",
          "category",
          "proficiency",
          "confidence",
          "source",
          "last_updated_at"
        ]
      },
      "EvidenceItem": {
        "type": "object",
        "description": "A single piece of evidence supporting a skill assessment",
        "properties": {
          "evidence_id": {
            "type": "integer",
            "description": "Evidence ID"
          },
          "evidence_type": {
            "type": "string",
            "description": "Type of evidence (e.g. certification, project, training)"
          },
          "title": {
            "type": "string",
            "description": "Evidence title"
          },
          "issuer_or_system": {
            "type": "string",
            "nullable": true,
            "description": "Issuing organization or system"
          },
          "evidence_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Date of the evidence"
          },
          "url_or_ref": {
            "type": "string",
            "nullable": true,
            "description": "URL or reference"
          },
          "signal_strength": {
            "type": "integer",
            "description": "Evidence signal strength (1-5)"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Additional notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the evidence was recorded"
          }
        },
        "required": [
          "evidence_id",
          "evidence_type",
          "title",
          "signal_strength",
          "created_at"
        ]
      },
      "EvidenceWithSkill": {
        "type": "object",
        "description": "Evidence item with associated skill information",
        "allOf": [
          {
            "$ref": "#/components/schemas/EvidenceItem"
          },
          {
            "type": "object",
            "properties": {
              "skill_id": {
                "type": "integer",
                "description": "Associated skill ID"
              },
              "skill_name": {
                "type": "string",
                "description": "Associated skill name"
              }
            },
            "required": [
              "skill_id",
              "skill_name"
            ]
          }
        ]
      },
      "EmployeeSkillProfile": {
        "type": "object",
        "description": "Complete skill profile for an employee",
        "properties": {
          "employee": {
            "$ref": "#/components/schemas/EmployeeRef"
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeSkillEntry"
            }
          },
          "total_skills": {
            "type": "integer",
            "description": "Total number of skills"
          }
        },
        "required": [
          "employee",
          "skills",
          "total_skills"
        ]
      },
      "SkillEvidenceResponse": {
        "type": "object",
        "description": "Evidence for a specific skill of an employee",
        "properties": {
          "employee": {
            "$ref": "#/components/schemas/EmployeeRef"
          },
          "skill_id": {
            "type": "integer"
          },
          "skill_name": {
            "type": "string"
          },
          "proficiency": {
            "type": "integer"
          },
          "confidence": {
            "type": "integer"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            }
          }
        },
        "required": [
          "employee",
          "skill_id",
          "skill_name",
          "proficiency",
          "confidence",
          "evidence"
        ]
      },
      "EmployeeTopSkills": {
        "type": "object",
        "description": "Employee's top/strongest skills",
        "properties": {
          "employee": {
            "$ref": "#/components/schemas/EmployeeRef"
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeSkillEntry"
            }
          },
          "limit": {
            "type": "integer",
            "description": "Limit applied"
          }
        },
        "required": [
          "employee",
          "skills",
          "limit"
        ]
      },
      "EmployeeEvidenceInventory": {
        "type": "object",
        "description": "All evidence across all skills for an employee",
        "properties": {
          "employee": {
            "$ref": "#/components/schemas/EmployeeRef"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceWithSkill"
            }
          },
          "total_items": {
            "type": "integer",
            "description": "Total evidence items"
          }
        },
        "required": [
          "employee",
          "evidence",
          "total_items"
        ]
      },
...
      "OrgRef": {
        "type": "object",
        "description": "Organization unit reference",
        "properties": {
          "org_id": {
            "type": "string",
            "description": "Org unit ID"
          },
          "org_name": {
            "type": "string",
            "description": "Organization name"
          },
          "business_unit": {
            "type": "string",
            "nullable": true,
            "description": "Business unit"
          }
        },
        "required": [
          "org_id",
          "org_name"
        ]
      },
      "OrgSkillSummaryEntry": {
        "type": "object",
        "description": "A skill summary entry for an org unit",
        "properties": {
          "skill_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "employee_count": {
            "type": "integer",
            "description": "Number of employees with this skill"
          },
          "avg_proficiency": {
            "type": "number",
            "format": "float",
            "description": "Average proficiency level"
          }
        },
        "required": [
          "skill_id",
          "name",
          "category",
          "employee_count",
          "avg_proficiency"
        ]
      },
      "OrgSkillSummaryResponse": {
        "type": "object",
        "description": "Org unit skill summary including child orgs",
        "properties": {
          "org": {
            "$ref": "#/components/schemas/OrgRef"
          },
          "includes_children": {
            "type": "boolean",
            "description": "Whether child org units are included"
          },
          "total_profiled_employees": {
            "type": "integer",
            "description": "Total employees with skill profiles in this org"
          },
          "top_skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrgSkillSummaryEntry"
            }
          }
        },
        "required": [
          "org",
          "includes_children",
          "total_profiled_employees",
          "top_skills"
        ]
      },
      "OrgSkillExpert": {
        "type": "object",
        "description": "A skill expert within an org unit",
        "properties": {
          "employee_id": {
            "type": "string"
          },
          "display_name": {
            "type": "string",
            "nullable": true
          },
          "job_title": {
            "type": "string",
            "nullable": true
          },
          "org_name": {
            "type": "string",
            "nullable": true
          },
          "proficiency": {
            "type": "integer"
          },
          "confidence": {
            "type": "integer"
          },
          "last_updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "employee_id",
          "proficiency",
          "confidence",
          "last_updated_at"
        ]
      },
      "OrgSkillExpertsResponse": {
        "type": "object",
        "description": "Skill experts within an org unit",
        "properties": {
          "org": {
            "$ref": "#/components/schemas/OrgRef"
          },
          "skill_id": {
            "type": "integer"
          },
          "skill_name": {
            "type": "string"
          },
          "min_proficiency": {
            "type": "integer"
          },
          "experts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrgSkillExpert"
            }
          },
          "total_matching": {
            "type": "integer"
          }
        },
        "required": [
          "org",
          "skill_id",
          "skill_name",
          "min_proficiency",
          "experts",
          "total_matching"
        ]
      }
    }
  }
}